Split html/converter.go into 11 focused files
Pure refactor — no behavior changes, no API changes. The 4367-line
converter.go is split into files by responsibility:

  converter.go           (734)  core types, ConvertFull, Convert, walkChildren
  converter_style.go    (1126)  CSS property application, transforms, counters
  converter_helpers.go   (603)  font resolution, text collection, gradients
  converter_block.go     (456)  div/blockquote/figure/definition list
  converter_table.go     (390)  table parsing and cell construction
  converter_paragraph.go (312)  paragraph, inline container, text runs
  converter_forms.go     (278)  input/select/textarea/fieldset rendering
  converter_image.go     (236)  image/SVG loading and embedding
  converter_flex.go      (234)  flexbox layout conversion
  converter_list.go       (85)  list and list item conversion
  converter_heading.go    (49)  heading conversion
  converter_link.go       (37)  anchor link conversion

All tests pass, lint clean.

Closes #34.
